This is a pull request to request merging changes that improve the performance of CSIDH. The changes have been tested and measured for the sibc module's test suite. Measurement was completed before and after applying this change set on on x86_64 as well as on POWER9/ppc64le to show the performance improvements.

Performance of CSIDH is probably improved on every architecture supported by sibc but I have only run the entire test suite to completion successfully only on POWER9 and x86_64. I don't have a sufficiently fast RISC-V system to run the full test suite in a single day and my prediction is that it would take a long time. I also did not benchmark this on arm64 or armv7 or i386. It may be worth measuring performance on RISC-V and arm64 but I don't think it should block this merge request.

The comparisons below are not entirely equal as some of the versions of software such as Python itself, and the underlying libraries are different. The CPU architectures are also very different and this means that the cross CPU comparison is largely interesting to show deficiencies in what seems to be general Python performance.

As an example to show the performance increase note the performance of one p512 test at random to benchmark before applying the patch.

On x86_64:
```
time pytest -k csidh_gae_p512_hvelu_wd1__unscaled
========================================= test session starts =========================================
platform linux -- Python 3.9.2, pytest-7.1.2, pluggy-1.0.0 -- /home/resilient/Documents/sibc/sibc-venv/bin/python3
cachedir: .pytest_cache
rootdir: /home/resilient/Documents/sibc, configfile: pytest.ini
plugins: parallel-0.1.1
collected 309 items / 308 deselected / 1 selected                                                     

test/test_csidh.py::csidh_gae_p512_hvelu_wd1__unscaled::test_group_action_with_random_exponents PASSED [100%]

================================= 1 passed, 308 deselected in 25.42s ==================================

real	0m25.646s
user	0m27.217s
sys	0m5.799s
```

On POWER9/ppc64le:
```
time pytest-3 -k csidh_gae_p512_hvelu_wd1__unscaled
========================================= test session starts =========================================
platform linux -- Python 3.10.4, pytest-6.2.5, py-1.10.0, pluggy-0.13.0 -- /usr/bin/python3
cachedir: .pytest_cache
rootdir: /home/user/src/sibc, configfile: pytest.ini
plugins: forked-1.4.0, xdist-2.5.0
collected 309 items / 308 deselected / 1 selected                                                     

test/test_csidh.py::csidh_gae_p512_hvelu_wd1__unscaled::test_group_action_with_random_exponents PASSED [100%]

============================ 1 passed, 308 deselected in 142.63s (0:02:22) ============================

real	2m23.509s
user	2m23.326s
sys	0m0.196s
```

After applying the patch, performance improved remarkably for this single test and additionally _all_ tests continue to pass.

On x86_64:
```
$ time pytest -k csidh_gae_p512_hvelu_wd1__unscaled
============================= test session starts ==============================
platform linux -- Python 3.9.2, pytest-7.1.2, pluggy-1.0.0 -- /home/resilient/Documents/sibc/sibc-venv/bin/python3
cachedir: .pytest_cache
rootdir: /home/resilient/Documents/sibc, configfile: pytest.ini
plugins: parallel-0.1.1
collected 309 items / 308 deselected / 1 selected                              

test/test_csidh.py::csidh_gae_p512_hvelu_wd1__unscaled::test_group_action_with_random_exponents PASSED [100%]

====================== 1 passed, 308 deselected in 18.73s ======================

real    0m18.954s
user    0m20.448s
sys    0m5.881s
```

On POWER9/ppc64le:
```
time pytest-3 -k csidh_gae_p512_hvelu_wd1__unscaled
========================================== test session starts ==========================================
platform linux -- Python 3.10.4, pytest-6.2.5, py-1.10.0, pluggy-0.13.0 -- /usr/bin/python3
cachedir: .pytest_cache
rootdir: /home/user/src/sibc, configfile: pytest.ini
plugins: forked-1.4.0, xdist-2.5.0
collected 309 items / 308 deselected / 1 selected                                                       

test/test_csidh.py::csidh_gae_p512_hvelu_wd1__unscaled::test_group_action_with_random_exponents PASSED [100%]

============================= 1 passed, 308 deselected in 87.69s (0:01:27) ==============================

real    1m28.410s
user    1m28.267s
sys     0m0.149s
```

The improvement for this single test on x86_64 is roughly seven seconds - not so impressive but certainly a big step in the right direction. On POWER9/ppc64le it is a reduction of nearly a full minute.

After applying the patch, the full test suite still takes a remarkably long time to run even when run on high end multi-core systems with many parallel tests. We probably want a way to run these tests quickly or people may be tempted to commit code without the full test suite running.

The full test suite POWER9/ppc64le _after_ the patch has been applied:
```
=================================== 309 passed in 28718.04s (7:58:38) ===================================
```

The full test suite on x86_64 _after_ the patch has been applied:
```
======================= 309 passed in 7259.67s (2:00:59) =======================
```


In summary, the relative CPU performance difference is a factor of roughly ~4.6 - ~4.7 for x86_64 vs ppc64le. Within each CPU architecture, the performance decrease in time from the previous revision of sibc is around ~26.3% for x86_64 and ~56.9% for POWER9/ppc64le, the absolute difference is ~30% for X86_64 and ~79.5% for POWER9/ppc64le.

I have not extensively profiled Python itself and I suspect that CPU vs CPU performance difference is in Python itself. This is part of the cost of nearly free portability by using pure Python. We may want to improve the performance with C, Golang, or Rust bindings at some point. It seems fruitful to consider improving Python itself such that everyone may realize generic improvements.

This patch greatly improves the performance of sibc in pure Python. I  request that it be merged if it passes a code review by @JJChiDguez. I am not the author of these improvements as the author prefers to remain nameless; I reviewed, and tested the patch extensively, and I am requesting a merge at the author's request.